requirements.txt
keras==2.10.0
matplotlib==3.6.2
mysql_connector_repackaged==0.3.1
numpy==1.23.4
opencv_python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
plotly==5.11.0
streamlit==1.14.0
streamlit_authenticator==0.1.5
tensorflow==2.10.0
tensorflow_gpu==2.10.0
